+++ lib/edgedb/protocol/codecs/config_memory.ex
@@ -41,9 +41,7 @@ defimpl EdgeDB.Protocol.Codec, for: EdgeDB.Protocol.Codecs.ConfigMemory do
 
   @impl EdgeDB.Protocol.Codec
   def encode(_codec, value, _codec_storage) do
-    raise EdgeDB.InvalidArgumentError.new(
-            "value can not be encoded as cfg::memory: #{inspect(value)}"
-          )
+    raise EdgeDB.InvalidArgumentError.new("value can not be encoded as cfg::memory: #{inspect(value)}")
   end
 
   @impl EdgeDB.Protocol.Codec