+++ lib/edgedb/protocol/codecs/float64.ex
@@ -51,9 +51,7 @@ defimpl EdgeDB.Protocol.Codec, for: EdgeDB.Protocol.Codecs.Float64 do
 
   @impl EdgeDB.Protocol.Codec
   def encode(_codec, value, _codec_storage) do
-    raise EdgeDB.InvalidArgumentError.new(
-            "value can not be encoded as std::float64: #{inspect(value)}"
-          )
+    raise EdgeDB.InvalidArgumentError.new("value can not be encoded as std::float64: #{inspect(value)}")
   end
 
   @impl EdgeDB.Protocol.Codec