+++ lib/edgedb/protocol/codecs/int32.ex
@@ -36,9 +36,7 @@ defimpl EdgeDB.Protocol.Codec, for: EdgeDB.Protocol.Codecs.Int32 do
 
   @impl EdgeDB.Protocol.Codec
   def encode(_codec, value, _codec_storage) do
-    raise EdgeDB.InvalidArgumentError.new(
-            "value can not be encoded as std::int32: #{inspect(value)}"
-          )
+    raise EdgeDB.InvalidArgumentError.new("value can not be encoded as std::int32: #{inspect(value)}")
   end
 
   @impl EdgeDB.Protocol.Codec