+++ lib/edgedb/protocol/codecs/named_tuple.ex
@@ -70,9 +70,7 @@ defimpl EdgeDB.Protocol.Codec, for: EdgeDB.Protocol.Codecs.NamedTuple do
 
   @impl Codec
   def encode(_codec, value, _codec_storage) do
-    raise EdgeDB.InvalidArgumentError.new(
-            "value can not be encoded as named tuple: #{inspect(value)}"
-          )
+    raise EdgeDB.InvalidArgumentError.new("value can not be encoded as named tuple: #{inspect(value)}")
   end
 
   @impl Codec