+++ test/support/scripts/setup-roles.sh
@@ -1,0 +1,5 @@
+#!/usr/bin/env bash
+
+set -e
+
+edgedb query --file test/support/scripts/edgeql/setup-roles.edgeql